requirements.txt
asgiref==3.3.4
Django==3.2
factory-boy==3.2.0
Faker==8.1.2
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
